This commit was generated by cvs2svn to compensate for changes in r25633, which included commits to RCS files with non-trunk default branches.

+++ Packages/Boost/include/boost/archive/iterators/wchar_from_mb.hpp
@@ -105,7 +105,7 @@ wchar_t wchar_from_mb<Base>::drain(){
     char buffer[9];
     char * bptr = buffer;
     char val;
-    for(std::size_t i = 0; i++ < MB_CUR_MAX;){
+    for(std::size_t i = 0; i++ < (unsigned)MB_CUR_MAX;){
         val = * this->base_reference();
         *bptr++ = val;
         int result = std::mblen(buffer, i);
